NER/requirements.txt
paddlenlp>=2.2.4
paddlepaddle>=2.2.2